Further update to allow delft3d z-layer communication files. The original code is a merge with the 20201104_delpar_zlayers code but based on the previous version of 20200109_kleissen_part.

+++ src/utils_lgpl/deltares_common/packages/deltares_common_c/src/shared_library_fortran_api.c
@@ -1,6 +1,6 @@
 //---- LGPL --------------------------------------------------------------------
 //
-// Copyright (C)  Stichting Deltares, 2011-2019.
+// Copyright (C)  Stichting Deltares, 2011-2020.
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Lesser General Public
@@ -111,7 +111,7 @@ void RemoveTrailingBlanks_dll(char * String)
  * ============================================================================
  */
 #if defined(WIN32) || defined (HAVE_CONFIG_H)
-long STDCALL OPEN_SHARED_LIBRARY(int64_t * sharedDLLHandle, char * library, long length_lib)
+long STDCALL OPEN_SHARED_LIBRARY(long long int * sharedDLLHandle, char * library, long length_lib)
 #elif defined (salford32)
 extern "C" OPEN_SHARED_LIBRARY(int64_t * sharedDLLHandle, char * library, long length_lib)
 #endif
@@ -136,7 +136,7 @@ extern "C" OPEN_SHARED_LIBRARY(int64_t * sharedDLLHandle, char * library, long l
     if (tmpSharedDLL->dllHandle != NULL)
     {
         error = 0;
-        *sharedDLLHandle = (int64_t) tmpSharedDLL;
+        *sharedDLLHandle = (long long int) tmpSharedDLL;
     }
 
     free(lib_name); lib_name = NULL;
